Task list is separated by Section headers defined in Contentful
A narly change was needed to `anyone_can_edit_their_answers_spec.rb` because we were trying to fake out a state where an answer existed and skipped the contentful stubbing. Because of this the journey has no section_steps and we can't use factories to create a factory with the right sections because the answers are needed up front. This route feels painful.

Instead the other option (which I went with) is to perform an update on the journey within the few tests where this happens to set the grouping after the faking of data.