Be more explicit about what links on journey map do
Previously, each step title would link directly to the step in
Contentful, with no context about the action which would happen. This
changes behaviour so that the step title is not a link, and there is an
explicit link to the "edit" action below.

This paves the way to adding additional context and information to each
step in the map.