Users can answer questions which expect a currency
The presenter adds the currency symbol and overloads the save method to
strip out commas.
Precision 11 has been set for the currency decimal
value within the database to ensure an element of future proofing. We
believe this value is higher than what any school would spend. This
helps give us confidence that there wouldn't need to be any changes to
the database in the future.

Co-authored-by: Tom <tomh@dxw.com>