Options are no longer capitalised in checkbox questions
The desired behaviour is that fields as displayed to the user in
checkbox questions are provided exactly as in Contentful. This commit
changes the behavior of `checkbox_options` in `StepHelper` to no longer
capitalise options.

NB: This does _not_ modify how checkbox options are rendered in
specification output, as further work is planned to provide a wider
range of options which will supersede this.