Fix section and step ordering in the task list
We discovered a bug on staging where the section _and_ step ordering was mixed up and not as the user defined in Contentful.

Our tests did attempt to cover this but perhaps without enough sections/steps and with different title names.

When we saved the `section_group` onto the journey as a hash it would set a correctly ordered hash into the database in a different order. When we retrieved it the order would be different.

To fix this I have been very explicit about the ordering we expect in the data structure we set on the journey. Whilst I could have used a simpler array of steps and trusted the natural array index to keep track of the ordering, this felt implicit and not clear to the reader.

We store the data structure with the correct order values and use a helper to unpack the ordering and get the step objects ready for the view to handle as it did before.