Contentful can mark a Step as hidden from the user by default
Steps that are marked as hidden will not appear in the task list. Meaning this change does not change anything for the user yet. The next change will look at at deciding when to mark a hidden step as ready or qualified to be shown.

This field is not required in Contentful for convienence. If the user doesn't specify Yes/No then we will assume the previous behaviour which is to always show it.