Fix bullet warnings by removing unnecessary scope
When adding a new feature this class starts to introduce bullet warnings due to steps being loaded that aren't used (because the scope was filtering them out)

If we remove this scope usage no unit tests fail which is a sign that it's not doing anything.

I added a new unit test to be sure of this. Making sure that we can safely ignore steps that don't have answers. This sits next to the test that makes sure we error when we find an unexpected answer.