Generic result object to support controller refactoring
So that we can perform more complex save logic outside of the controller we need to pass a result back to the controller. This object needs to communicate success and failure to support redirections as well as the original object which may or may not have errors added to it.