Answering a step can show the user additional questions
Some resilience against answer casing is included to be match 2 strings of any different casings together as a match.

Since the `show_additional_step` value will contain a `contentful_id` and there may be many steps with the same value, we have to scope the search by only the journey the user is in. We do allow this value to be injected straight into a `find_by` query which is usually cause for concern however Contentful is a trusted source of input.

There is more work to do to ensure that if a shown/hidden step also includes logic to show/hide more questions, that we cascade this update all the way down.