Hiding an optional step will hide all connected hidden steps
We know traversing the tree like this is not efficient. Ideally we would be keeping track over the tree of connected steps and be able to hide all without N+1 queries. We are not aware of real use cases that will chain many of these steps together right now, we imagie there may be at most 2-5 connected together.

When an optional step is hidden again we hide all of it's dependent steps without reevaluating. The dependant steps will become visible with their previous qualifiying answers if the user changes their mind again.