Date validations
When a date cannot be parsed into a valid date object within Ruby, we set it to nil and pass it onto the activerecord `presence` valdiation.

This does not feel like the correct approach. Ideally Rails will recieve the parameters `response_dd`, `response_mm`, `response_yyyy`, coerce them into a single attribute and run validation on the combination. This is how the MOJ date helper works [1]. Given that three other members of the team have already spent a lot of time trying to find the right solution, this approach is very much "just get something working". This approach is lifted from BEIS RODA [2].

We include a locale file for activerecord translations so we can change the message from "Can't be blank" to something more generic. Not ideal because it doesn't say what was wrong but it does at least prevent the critical error and instruct the user in some way so that they can recover.

[1] - https://github.com/ministryofjustice/gov_uk_date_fields#3-tell-your-models-attributes-are-gov_uk_dates
[2] - https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/blob/f94d7fdb18dcf8df15f8d3a36de7e9a67e2e822a/app/controllers/staff/activity_forms_controller.rb#L81