Add markdown parsing to question help text
This commit adds a new `help_text_html` method to a `Step`, which passes
the `helpText` field from Contentful through a markdown parser.

Using this field when presenting the help text to the user allows for
richer formatting to be used, including (most pertinently) paragraph
breaks and unordered lists, although the full range of markdown
formatting is supported.

Changing the behaviour of `hint` within the form builder to use a lambda
instead of a further hash means that the text will be wrapped in a
`<div>` instead of a `<span>`, which keeps semantics intact.