Extended checkbox pattern can store multiple pieces of further information
The DfE Form builder gem was only designed to support a single checkbox option that would show a text field, this is called "other", and allows the user to provide a single choice.

We are pushing this pattern to work for our own needs: every checkbox can have a field the user can give further information in.

This is difficult. The formbuilder gem expects and requires an activerecord object to behave normally. This object needs to have a unique attribute for each "further information".

We can't add `n` fields to the database table for `1_further_information`…`n_further_information` because we can't ever know ahead how many options will exist.

This approach modifies the form object when displayed to the user. It dynamically adds ephemeral attributes to the object so that the form object can work with unique fields such as `yes_further_information` rather than only `further_information`.