Multiple optional questions can be shown and hidden
Before this change we allowed the content user to "bolt on" one additional question if the answer matched an expected pattern.

This change allows them to show as many additional questions as they like. Though the current need is for 2 additonal questions. We do this by accepting an array of identifiers rather than a single string.

The amount of calculation required to determine and follow the question branches through to their resolution is not ideal and we know this. A better approach would be to keep track of the state of the tree of questions and make the required database updates with fewer database queries. This would require more time than we have in the last sprint so we are opting for something that works and accepting the technical debt will need to be repaid.