Ignore weak Brakeman warning on XSS
This alert is warning us about this line `@specification_html.html_safe`

The `specification_html` is being sourced from Contentful and is indeed being rendered to the page directly without any escaping or sanitising.

I am ignoring this alert as we trust Contentful to feed data directly into the serivce. There is a risk that the login for Contentful could be compromised and the specification template could be made to render malicious code to the user.

This ignore file was generated automatically using the `brakeman -I` flag.