Toggling additional questions checks multiple rules
Before this change there could only be a single additional question rule. Allowing the user to specify at most one question to show based on at most one matching answer.

This changes extends the existing logic to allow forking. Now we can define multiple rules which allow the user to show _either_ question A or question B based on an existing answer.

We have to continue to ensure that all further connected steps to A or B are recursively shown or hidden as per the change in answer.

We start to pass around an array of step objects rather than ids as we now need to check each 'next_step_id' to make sure it qualifies as a matching next step.