Move rendering the specification to a service object
Previously, all rendering work for the specification was done in the
journey controller. This moves the job of combining the template and the
answers into a rendered output into a new service object, which gives us
clearer separation of what is responsible for the rendering.

This paves the way for making the docx output different to the web
output, for example by adding warning messages.

Updates Brakeman ignore file to account for the .html_safe operation now
happening in the service object.