Radio answer responses no longer include further info
We are changing the way we are making the `further_information` data available to the specification template.

Instead of automatically interpolating the `response` together with the `further_information` (separated by a hyphen), we are going to make them both available to the Liquid template as individual variables and allow the content team to choose how they are presented.

The first step of this work is to make the response field for Radio answers exclude it's further information.