Specification can choose to show radio further information
The choice of an array to house this hash was made because the same pattern will be used to deliver checkbox answers, of which there may be multiple further_information values to consider.

Some questions don't have a `further_information` field so we have to guard against the other types.