Fix page jump for radio questions without further info
When the radio question does NOT have a further_information of any sorts then when the user selects an option, the radio elements will jump slightly on the page.

This is caused by a new HTML element being added by the gem when `f.govuk_radio_button` is passed an empty block. In order to avoid this jump, we default back to the old (slightly repetitive) version of the code that has an else condition that uses this method without passing any block. This fixes the issue.

>  block (Block) — Any supplied HTML will be wrapped in a conditional container and only revealed when the radio button is picked [1]

[1] https://www.rubydoc.info/gems/govuk_design_system_formbuilder/GOVUKDesignSystemFormBuilder/Builder#govuk_radio_button-instance_method