Checkboxes can have their options separated by an OR
Though the user need was for this to only work with radio questions, the mechanism to provide this for radios is the same as checkboxes. The same Contentful `options` field is used and we provide guidance in the form of help text within contentful as to what attributes are useable. Given this is easily supported now, we add this for checkboxes too as a nice to have.