Steps can be marked as skippable
We need the ability to turn off mandatory validation in order to allow a user to explicitly answer a question by choosing none of the answers. This marks the question as "complete" and distinguishes it from a question that simply hasn't been answered yet and which requires attention.

We are going to use the presence or omission of text to infer if it is skippable or not. A missing skip text means it isn't a skippable question. This removes the need for 2 fields, one as a boolean and another for the text.