Checkbox answers are only required if the parent step isn't skippable
Users may need to explicitly state that their answer is "none of the above". We are going to do this by adding a skip button that ignores any checkbox selections, this change facilitates that.

With the radio pattern (given only 1 choice can be made) an additional "Other" option can be presented naturally.