Checkbox answers can be completed without selecting a given option
The content team can control whether a question is skippable by providing button content. If found, we will show an additional secondary button at the bottom of the list of options.

This could not be a link as it would always be a GET request, rather than POST.

If a skip occurs we forcefully update the answer with an empty set of params, undoing any previous choices in the case of an update.

We could extend this to all question types but it isn't needed right now and checkboxes is the only skippable question type. In future, should this be needed we could push the responsibility down into SaveAnswer after tidying up how we are sending different parameters for the different `params` too.