Rename existing use of "planning" to "specifying"
When we started the service we weren't sure what the name of the journey we were building was. We assumed it was planning however this quickly became the "build a spec" journey and we didn't come back and rename.

Now we are adding a planning start page, we are renaming the existing code to make this change distinguishable from the existing language/names.