Checkbox answers remember if they've been skipped
Currently we infer "completed" but which questions have an associated answer object. We can be more explicit about this for readability and future scoping.

Now that answers know if they are being marked as skipped, we can use
this information to turn validation back on for one case. When a user
doesn't choose a checkbox option OR the skip button, we _still_ want
them to make one of those two selections. We can now show a validation
error when this happens.