A checkbox answer is still required unless it is skipped
The user must now make a choice between one of the options and clicking the skip button. They can no longer click "continue" or "update", they must make a decision.

The form error doesn't wrap all the way down to the "or" and skip button form group. This would be nice to have but I can't see a way to easily do this without hacking the form builder gem to wrap multiple form groups if present. Hopefully this will work for the few cases where skipping is needed.