Checkbox answers are not automatically joined by a comma
Instead of performing this operation in Rails, we leave the presentational logic to Contentful.

The same logic can be achieved by defining the following in the `specification_template`:

```
{{ options | join: ", " }}
```

This enables the user to present the options in different ways. Not only as a single paragraph but as a list of options.