Checkbox answers are give to the spec builder as a hash
Instead of only returning a string response for Liquid to use for each answer, we extend the functionality by returning a hash instead.

We continue to include the stringified value and we start to extend by adding new fields like `skipped` so the content team can determine if an answer is blank OR was skipped in an explicit way.

We have many hops to migrate each presenter so the `to_param` method has been added to each presenter and each will take a similar approach until all answers return a hash.

Once this migration has been completed the `extended_answer_x` variables will be made redundant and removed. To keep tests green for now, they remain.