Checkbox answers return a concatenated string of options
Given that the Liquid `contain` method cannot be called on a value that is of type Array, we decide to do extra lifting on the server side to reduce the logic needed in Contentful.

https://shopify.github.io/liquid/basics/operators/

Before this change an additional `assign` and `join` would be needed before any `contains` could be used. This is a lot for the user to remember if they want to present a comma separated string instead of iterating through an array of options, if they wanted to present a list:

```
  {% assign selections = answer_7p7URWwZbkjqqzo3E0k1v4 | join: ", " %}
  {% if selections contains "extra" %}
    <p>when there is an answer but no selection was made, this should show</p>
   {% endif %}
```

We expose a new option called `concatenated_response` in our new hash of options.