Remove @extended_answer variables used by the spec
Now that the @answer variables have a has structure with all the further information included, these aren't needed.

Use hash with_indifference_access so that Liquid can use the `["key"]` syntax. Using symbols `[:key]` breaks the liquid parser.

In the fixture we include a specification_template that includes a subtle difference. We are using both syntaxes for accessing the hash. The more traditional `["key"]` and the dot syntax `.key`, just to ensure we have test coverage for both.