Multiple answers can unlock the same branching question
This change fixes a bug where one question with branching logic could not unlock the same question baed on more than one required answer.

For example, if question A should show question B if the answer was "Blue" **or** question B if the answer was "Red". The rules would be iterated through one at a time, the first rule would correctly unlock the first question, but the second rule would incorrectly hide it again (straight away). The behaviour we want is: if any rule matches we show it and it stays shown until updated again.

We can achieve this by simply switching the order of how we show and hide all rules. If we go through all the hides first we can never have the case where we re-hide a question that was shown in the current toggle action.