Fix skipping answers so that previous choices are removed
The previous version of this code `answer.assign_attributes({:response=>nil})` does not change the response column. Probably because at a database level a column with the type Array cannot be set to nil so the update is ignored.

Setting this to an emprt array `[]` has the same result, nothing changes.

The only way I found to remove the old choices is to set an empty string in the array.