Branches can be hidden when the node includes multiple rules
Before this change any branches that had child nodes with had a _single_ rule worked fine.

When any child node had _more than one_ rule then we got a `TypeError: can't cast Array` [1].

This was happening because `journey_steps.where(contentful_id: [["8"]])` returns an object called `Step::ActiveRecord_AssociationRelation` which is automatically presented as an array of objects.

The problem comes when a second ID is passed and `flatten` isn't used: `journey_steps.where(contentful_id: [["8"], ["9"]])`. This code only returns a `Step::ActiveRecord_AssociationRelation` object and doesn't present itself as an array of steps. This fails to query for steps but doesn't explode. We see the problem when we then call `.update` upon it.

[1] https://rollbar.com/dxw/dfe-buy-for-your-school/items/91/?item_page=0&item_count=100&#traceback