Further info for variables with special chars is available to the spec
Looks like the application of the `parameterize` method didn't touch enough places and our test suite had holes in it for keys that had special characters.

The test case covers an exact case from live.

We reuse the `AnswerHelper` and this time check that `parameterize` is no longer defined in the rest of the code base.