Specification document includes an incomplete warning
The way this feature is tested is not the preferred option but it's the only option I could make work given I don't think I can download the word file and open it for inspection with Capybara.

Option 1 would be as above, we download the actual file and inspect the contents. This doesn't work. The closest I can get is to inspect the `page.body` and I get the whole structure of the word doc which is built using XML in _binary_. Attempts to `unpack` this failed. Eg: ,`"PK\x03\x04\x14\x00\x00\x00\b\x00\xD8^oR\xDFB\xF3w\x84\x01\x00\x00\x7F\x06\x00\x00\x13\x00\x00\x00[Content_Types].xml`

Option 2 would be to assert that the controller calls `render` with the `content` value that is expected. This would be close enough to assert that Rails was at least attempting to render the right content, via the htmltoword gem behind the scenes. The following successully asserts that the right value is passed but then fails as render is recieved twice, once with the args and then again with the binary blob. Even with `and_call_original`:

```
expect_any_instance_of(SpecificationsController)
.to receive(:render)
.with(hash_including(content: "<article id='specification'><h1>Liquid </h1></article>"))
.and_call_original
```

Option 3 is what we have. We assert that the gem that creates the content receives create with the expected HTML. What's not ideal is that there is no assertion that it's the response that's rendered to the user. We haven't got complete coverage and confidence isn't high.