RadioAnswers can store multiple further info and present only..
.. the one that's related to the users choice

We use the existing pattern for checkboxes to store the further_information as a jsonb value. This allows us to manage it as a hash. We apply the same AnswerHelpers to set and read the dynamic further_informatiion fields.

This choice will also store further_information values given to other radio options should the user change their mind. Whilst this is not the need, we see some small benefit to being able to replay back past values for convenience. The presenter plays a key part in only presenting the chosen option to the specification.

Co-authored-by: Nick Jackson <nick@dxw.com>