Refactor by moving feature tests that only users should see
Before we make any code changes to restrict this behaviour to only school buying professionals we reorganise these test files.

The file names no longer need to be prefixed with "anyone_" now a folder is used. The also all had "can" which can be removed so we're left with a verb to start the name of each test which seems more readable.