Fix checkbox answer to save choices after previously skipping
The skipped value was being set to true when a checkbox question was
skipped, but the opposite case didn't set it to false. This meant that
once a question had been skipped once it would never become not skipped.

This commit adds setting skipped to false in further_information_params
for checkbox questions.