Users can sign out of the service (and DSI)
Another term for this in the OIDC standard is "Global Logout - OP Initiated Logout with Back-Channel Communication".

When the user signs out of our service:
1. logout from the app is triggered
2. we call the app logout mechanism on the end users behalf
3. we destroy the local session
4. we call the logout URL given to us for this user within DSI
5. DSI should then decide whether to trigger the logout process for the other Relying Parties

Digestable breakdown of the spec: https://medium.com/@robert.broeckelmann/openid-connect-logout-eccc73df758f
Official spec: https://openid.bitbucket.io/connect/openid-connect-backchannel-1_0.html

We ignore a Brakeman warning after reviewing it. It is true the session
is passed to an object that is then used to construct a redirect url:

```
UserSession.new(session: session).sign_out_url
```

I think this is OK to ignore because the method `sign_out_url` only
allows 2 hardcoded paths to be calculated. If the
`dfe_sign_in_sign_out_token` is present then a payload with
`auth_dfe_signout_url` is created. If not, the root_path is used. There
is no case where session data is used to decide the redirect location.