Flag to use Omniauth in dev, instead of DSI
Inspired by this work on Apply [1]

The flow takes you to a form where you can fill in your desired DSI UID
and email, which will be returned to the app as though via a DfE Sign-in
authentication payload.

Note the :fake_callback method in the SessionsController which is just
an alias of :callback. This is necessary because the Omniauth
`development` strategy we're using mimics some real-world authentication
providers in that it POSTs back to our app. Because the POST has come
from "off-site", no CSRF token is expected.

[1] https://github.com/DFE-Digital/apply-for-teacher-training/commit/3ce79adf194378eda2e788dfff9e095ba841881f#diff-0a10c197c923aa39f522070b424c123d1a93ae9f2e36f662db5177ec617b0386L7