Users are redirected through DfE Sign in when they click Start
This work is inspired by the work done on other DfE projects. Namely Teaching Vacancies (which has since moved to Devise) and Apply for teacher training [1]

In the data we request from DSI we only ask for the `openid` scope. This means we do not receive the names, email address or any other information. This should be okay as we don't think we need any personally identifiable information from our users.

For the failing test case we want to check that if the authentication proccess fails that we present something to the user. This is unlikely to happen under normal circumstances as if a user gets their credentials wrong they'll not be redirected. This _should_ cover the case where we have misconfigured something so Rollbar is sent an error to warn the team of a potential misconfiguration.

There is a subtle oddity/difference that was neceassary to get this to work. Specifically in the callback phase of OIDC which fails with "invalid client" if we do nothing. In the Omniauth initialiser we include an additional option `client_auth_method: :client_secret_post`. This is needed so that when Rack OAuth2 (a dependency of the omniauth_openid_connect gem) gets to this point [2] then we fall into the `else` statement and add our credentials as parameters. The basic auth (:basic) doesn't work as a default for some reason and DSI gives no feedback or logging as to why this might be the case. The symbol `client_secret_post` relates to the (only) auth option selected in the DSI console in service settings, but in practice it would work with any value not present in that case statement.

[1] https://github.com/DFE-Digital/apply-for-teacher-training/blob/master/config/initializers/omniauth.rb
[2] https://github.com/nov/rack-oauth2/blob/master/lib/rack/oauth2/client.rb#L84