An unrecognised DfE Sign-in UID will create a new record
Users won't be invited to our service but we can infer that a successful log in from DSI is enough to generate a new user record. This user record will then be found the next time they sign in, rather than constantly creating new user records. This should allow us to connect other records like journeys to it with integrity.