Most existing endpoints require authentication
Public pages like the start pages, the health check and error rendering does not require authentication. Crucially the sign in end points don't require authentication either, or nobody would be able to get into a signed in state.

Use the new helper `user_starts_the_journey` to capture the steps needed to start the journey into a single method that could be chained. This has been applied across the test suite to give us a single definition. It's imagined we'll need to extend this pretty quickly as the definition of starting changes after the introduction of a dashboard.

The existing feature tests use a new helper `user_is_signed_in` to quickly stub a valid session. This keeps them fast and requiring minimum alerations to support the new authentication logic that would come before every feature test. With so many tests, making each go through the Omniauth flow and the sign in controllers would amount to a lot of wasted time. The test coverage for signing in is covered in a dedicated feature and request spec.