New public endpoint to invalidate cached contentful entries
This endpoint exists so Contentful can send webhooks on update events. With this information we can make targetted cache invalidations, enabling the content team to see their changes published quickly, with a small feedback loop. The preview environment is not useful to test branching logic so the team are using staging for this type of testing.

Given the service takes a shallow copy of all entries when the user starts a journey, all existing journeys and their steps shouldn't be at risk of breaking should the cached entry is replaced with an invalid one. An invalid entry would look like one with invalid JSON within it's configuration for the advanced fields. We should probably validate entries before setting them to the cache to remove this risk, this risk is not new. Any changes to the Contentful Master environment could break journeys.

Next steps are to authenticate the endpoint so that the public can't make infinite requests to our Redis.del function. Whilst the impact of finding a valid entry ID and deleting it are minimal to none, we may become open to a denial of service against our Redis database.