Docker build succeeds without needing environment variables
APPLICATION_URL was recently added and as an initialiser causes the docker build process to fail.

There is no way to provide a real URL during the build stage since we don't know what environment the docker image will be used in. Instead we guard against this variable being unset and if it is, we do nothing, reverting back to Rails defaults.

When the container is started off this image, the initialisers should re run and have the environment variables available to set this as originally intended.