Debug: are environment variables set correctly?
GPaaS deploys from GitHub Actions are currently failing. When deploying the apps locally with existing environment variables, it works.

The error is an odd one to read. It looks like bundle is called to prepare the database and crashes. It seems that we might not have an RAILS_ENV, meaning no environment is loaded.

When deploying the app from GitHub Actions, it fails.

We can't read the secrets set in GitHub to verify they are correct, though I have manually renentered each one to no affect.

This change assumes there may be a problem in the script that converts the yml file to environment variables.