Add Sections as a model in the application
Currently, only Journeys & Steps are models in the application, and Sections
are stored as a json blob `section_groups` on Journeys.

We want to add a new model type to Contentful & the application between Section
and Step ("Tasks") and to do so it makes sense to add Sections as a concrete
model in the application, and later add Tasks as a model as well.

This PR is the first step in the process. Sections are added as a model,
attached to a Journey. Eventually we will add tasks as a `has_many` association
to Sections, and Steps as a `has_many` association to Tasks.