Sign in helper will take every user past the root path
Instead of starting on the dashboard_path, we follow the real user journey of hitting our landing page first.

Before this, visiting the dashbaord would redirect the user to auth/dfe with a GET request. This needed to change due to a security vulnerability to become a POST request. This redirection then becomes impossible. Despite the vulnerability, changing to visit the root_path first in most feature tests makes sense for complete end-to-end coverage.