Resolve CSRF vulnerability with Omniauth
We change the link_to to a button_to to change the sign in button from making a GET request to a POST request. The POST request will be protected by Rails CSRF protection to prevent malicious requests from signing in when they shouldn't.

https://github.com/omniauth/omniauth/wiki/Resolving-CVE-2015-9284

Remove now unused sessions resource and redirect/link to `/auth/dfe`directly.

If a user becomes unauthenticated we change where they go. Instead of automatically taking them to the sign in page - through a GET request - we link them back to the root_path so that they may click "Sign-in" with the safe POST. This makes it safer, though slightly more inconvienent for the user. We could add flash messages in future.