Add noindex,nofollow meta tag to all pages
As per Gov.uk guidelines, all pages on the application should have a "robots"
meta tag with the content "noindex,nofollow"

https://www.gov.uk/service-manual/technology/get-a-domain-name

This does not require a robots.txt entry. A robots.txt entry will prevent the
crawler from seeing the noindex directive, so pages on the service may still
appear in search results.