Refactor how answer params are passed to the SaveAnswer service
Having to account for 3 different optional parameters within the service was already requiring us to be overly verbose with a case statement to handle things.

Interestingly I found that `date_params` is always present `{response: nil}` - just one bug that makes it harder to extend this class.

Move out the responsibility of determining which params to save up into the controller.