StringSanitiser service to strip unsage characters all answers
We need a service to make strings safe before we persist to the database. Doing so removes the risk that we render any malicious scripts back to the users of the service.

Usually Rails protects us by stripping these from our views for us. We can get around this by using html_safe.

We are doing something a little unusual that requires us to use html_safe. The HTML within the liquid_template is to be trusted, but the user input should not be. Before this change we are combining both sources of HTML and marking them as html_safe. This introduces the possibility of XSS.

At the moment all users are required to sign in and users cannot see eachothers content within the servce. This mitigates the risk to users being able to XSS themselves. However, this may not always be the case and a future change (perhaps spec sharing) could change the situation without an obvious sign. It's safer for all if we sanitise this data on the way into the service.

We create a new class instead of a private method within SaveAnswer in order to get better test coverage on what is a little complex given the different data structures `params` can take.