Refactor cache warming exception handling for clarity
Make it clearer which code we are expecting to handle exceptions from. On reading this class again after some time it first appeared like the exception was being thrown from within `store_in_cache` which would have meant some entries were cached while the rest were all extended, regardless of their state. This was wrong so this redesign hopes to avoid this confusion.