Sanitise user answers on the way out
Though we are sanitising on the way in we need to be thorough and ensure that we are also making sure user input is safe on the way out.

This is defence in depth but could happen for a number of reasons:

1. The sanitisation logic for data on the way in could break or be bypassed by new logic
2. Data could be added to the database by another mechanism such as rake task, directly from contentful or a new form in the service
3. Existing data that exists in a real database may already have unsafe data in it.

There was a test on GetAnswers that checked we were returning copy with <p> tags which informed us of an oversight. LongText answers convert user input to HTML so we can render new lines. This change also adds p, b and i tags to the sanitisation whitelist.

Changing from the `FullSanitizer` also changes the behaviour for some of our tests. Instead of stripping the conents within a script tag, it now strips ONLY the tags and returns the contents. This is still safe and if a person wants to hack the form I don't think we should be concerned with going out of our way to tidy their input for them.