Style specification sections that use a class
There only used to be one specification element but since we are splitting this across multiple fields we can expect multiple sections. It would be right to stop using `id=` and start using `class=`.

Continue to support both to keep it backwards compatible as using more than 1 ID on the page isn't good but it is functional.

In future we can remove the `#specification` when the existing definitions have been switched over.