Create Steps within Tasks when creating a new Journey
When creating a new Journey, we want to nest the Steps within Tasks.

To enable the tests to pass without updating all of the existing fixtures to
have Tasks encapsulating all Steps, we use login in CreateJourneyService to see
if the (faked) response from Contentful contains tasks, if not we create a
"fake" task which only contains each step. Later on we will use login in the
view to show these steps as direct children of the Section, without an
encapsulating task