Fix fixtures to align with Contentful for multiple spec fields
If you ask for a category over the Contentful API you get `specificationTemplate` rather than `specification_template`. Fortunantely this hasn't caused a problem so far as Contentful makes the method `specification_template` automatically available as a getter on `Contentful::Model`. This change makes it more accurate to reduce the chances that this causes a problem in future.

This was noticed because Contentful::Model behaved unexpectedly for `specificationTemplatePart2`. Instead of making `specification_template_part_2` available, it removed the last underscore and made `specification_template_part2`. The numbers are a key part to making we sort the fields correctly so we can't change to `specification_template_part_two`.