Enable Rails Content Security Policy
We don't want scripts or images being loaded from external sites.

Following the principle defence in depth this could be possible if user data was not correctly sanitised. They could inject a `<img src="evil.com">` into the client that would enable malicious activity.

Turning this policy on tells clients not to try and load content from anywhere that's not the current domain(self) and HTTPS.