Prevent users from having more than 1 active session
If a second authentication for a unique DfE Sign in UID happens, we remove the session data for the previous session, effectively signing them out.

On the first users next request they will be taken to the root_path with a flash message that says they've been signed out. The message is generic because we have no current way to connect the new session to the old session. We can't say "Someone else logged you out" without more work that might see us storing more state in redis.

We can't configure the Redis session store keys. They are in the format `2::28734342427213406823417860`. This long value corresponds to the private_id of the Session. As we can't manipulate this key to include the DfE Sign-in ID: `2::28734342427213406823417860:<dfe-sign-in-id>` we have to keep track of who is logged in and what where their session data lives. We d this using `RedisSessionLookup`.