Show a list of steps on the task page
In a previous commit we added a task#show page which was empty. This commit
adds the task's steps to this page, using the same UI pattern that was
previously used on the journey#show page.

Also, when a user answers a question from the task#show page, they will be
redirected back to the task#show page. If they answer a question from the
journey#show page, they will be redirected to the journey#show page