Add a helper method to Task to indicate if it has only one step
Task#has_one_visible_step? will be used to redirect the user appropriately
when they answer a question.

If the question belongs to a task with only one step, then the user will be
redirected back to the journey show page.

If the question belongs to a task with more than one step, the user will be
redirected back to the task show page.