Docker build works when RAILS_ENV is production
When Redis was added as a session store CI successfully built and tested the image.

When merging this - and all subsequent work - into `develop` the build failed https://github.com/DFE-Digital/buy-for-your-school/actions/runs/773887911

This happened because CI builds docker images with RAILS_ENV=test (for testing) and on deploy to devleop RAILS_ENV=production is used. This causes this asset:precompile to run for the first time https://github.com/DFE-Digital/buy-for-your-school/blob/develop/Dockerfile#L66.

Guard against this error by providing a default. The app just needs to start, it doesn't need to actually connect to redis.