Spec files are copied into the Docker image so Rspec runs them
Before this change CI was silently passing because rspec was not running any tests since no files were available to it.

The longer term approach would be to use a multi stage docker file as seen in this Rails Template proposal [1]. For now though, we need CI to run our tests.

[1] https://github.com/dxw/rails-template/pull/211/commits/5966236a533f07d330a5d24cb9ecc9d6fd52aa89#diff-dd2c0eb6ea5cfc6c4bd4eac30934e2d5746747af48fef6da689e85b752f39557R112