Align script/test so it's closer with scripts to rule them all
When we created this project the Rails Template didn't have the Scripts to Rule Them All merged in. Though the current version of script/test on Rails Template [1] includes many more things we would have to change our Dockerfile. This change adjusts it so we reach parity + run standardrb.

CI added to docker-compose.ci.yml so we don't run linting in auto fix mode.

[1] https://github.com/dxw/rails-template/blob/develop/script/test